# src/main/resources/archetype-resources/contract/src/main/resources/openapi/openapi.json
{
  "components" : {
    "schemas" : {
      "MyTableRequest" : {
        "properties" : {
          "description" : {
            "description" : "A description",
            "example" : "Example.",
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          },
          "myTableParentId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "description" : "A name",
            "example" : "Juan",
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          },
          "surname" : {
            "description" : "A surname",
            "example" : "Sanchez",
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "name" ],
        "type" : "object"
      },
      "MyTableResponse" : {
        "description" : "Datos de un registro.",
        "properties" : {
          "description" : {
            "description" : "descripción libre",
            "type" : "string"
          },
          "myTableParentId" : {
            "description" : "identificador del registro padre, si lo tiene",
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "description" : "nombre",
            "type" : "string"
          },
          "surname" : {
            "description" : "apellidos",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "PageMyTableResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/MyTableResponse"
            },
            "type" : "array"
          },
          "empty" : {
            "type" : "boolean"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "numberOfElements" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "PageableObject" : {
        "properties" : {
          "offset" : {
            "format" : "int64",
            "type" : "integer"
          },
          "pageNumber" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageSize" : {
            "format" : "int32",
            "type" : "integer"
          },
          "paged" : {
            "type" : "boolean"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "SimpleApiResponse" : {
        "description" : "Respuesta simple con un único mensaje.",
        "properties" : {
          "message" : {
            "description" : "texto del mensaje",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "SortObject" : {
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "title" : "${artifactId} API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/v1/examples" : {
      "get" : {
        "description" : "Lista todos los registros.",
        "operationId" : "listAll",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/MyTableResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "los registros existentes; lista vacía si no hay ninguno"
          }
        },
        "summary" : "Lista todos los registros.",
        "tags" : [ "example-controller" ]
      },
      "post" : {
        "description" : "Crea un registro nuevo.\n\n <p>La respuesta incluye la cabecera <code>Location</code> con la URL del recurso creado.</p>",
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MyTableRequest"
              }
            }
          },
          "description" : "datos del registro a crear",
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "format" : "int64",
                  "type" : "integer"
                }
              }
            },
            "description" : "el identificador asignado"
          }
        },
        "summary" : "Crea un registro nuevo",
        "tags" : [ "example-controller" ]
      }
    },
    "/api/v1/examples/hello" : {
      "get" : {
        "description" : "Devuelve un saludo en texto plano.",
        "operationId" : "sayHello",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "el texto <code>Hello World</code>"
          }
        },
        "summary" : "Devuelve un saludo en texto plano.",
        "tags" : [ "example-controller" ]
      }
    },
    "/api/v1/examples/helloDto" : {
      "get" : {
        "description" : "Devuelve un saludo como objeto JSON.",
        "operationId" : "sayHelloDto",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SimpleApiResponse"
                }
              }
            },
            "description" : "un mensaje de saludo"
          }
        },
        "summary" : "Devuelve un saludo como objeto JSON.",
        "tags" : [ "example-controller" ]
      }
    },
    "/api/v1/examples/paginated" : {
      "get" : {
        "description" : "Lista los registros de forma paginada.\n\n <p>Admite ordenación por cualquier campo con <code>sort=campo,dirección</code>, y el parámetro puede\n repetirse para ordenar por varios criterios.</p>",
        "operationId" : "listPaginated",
        "parameters" : [ {
          "description" : "Numero de pagina (empezando en 0)",
          "in" : "query",
          "name" : "page",
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "Numero de elementos por pagina",
          "in" : "query",
          "name" : "size",
          "schema" : {
            "default" : 10,
            "type" : "integer"
          }
        }, {
          "description" : "Orden: campo,direccion. Puede repetirse",
          "in" : "query",
          "name" : "sort",
          "schema" : {
            "default" : "id,asc",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageMyTableResponse"
                }
              }
            },
            "description" : "una página de registros con sus metadatos de paginación"
          }
        },
        "summary" : "Lista los registros de forma paginada",
        "tags" : [ "example-controller" ]
      }
    },
    "/api/v1/examples/{id}" : {
      "delete" : {
        "description" : "Elimina un registro.",
        "operationId" : "delete",
        "parameters" : [ {
          "description" : "identificador del registro a eliminar",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "summary" : "Elimina un registro.",
        "tags" : [ "example-controller" ]
      },
      "get" : {
        "description" : "Busca un registro por su identificador.",
        "operationId" : "getOne",
        "parameters" : [ {
          "description" : "identificador del registro",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MyTableResponse"
                }
              }
            },
            "description" : "el registro encontrado"
          }
        },
        "summary" : "Busca un registro por su identificador.",
        "tags" : [ "example-controller" ]
      },
      "put" : {
        "description" : "Actualiza un registro existente.",
        "operationId" : "update",
        "parameters" : [ {
          "description" : "identificador del registro a actualizar",
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MyTableRequest"
              }
            }
          },
          "description" : "nuevos datos del registro",
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "summary" : "Actualiza un registro existente.",
        "tags" : [ "example-controller" ]
      }
    }
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "url" : "/"
  } ],
  "tags" : [ {
    "description" : "Operaciones de ejemplo.\n\n <p>Todas las rutas exigen un token válido en la cabecera <code>Authorization</code>, salvo las sondas de\n salud. Las respuestas de error siguen el formato <code>application/problem+json</code>.</p>",
    "name" : "example-controller"
  } ]
}